Full Rewrite
Including:
* MacOS support
* Seed saving
* Crash fixes
