+++ .github/workflows/build.yml
@@ -15,8 +15,8 @@ jobs:
         - name: Windows
           os: windows-latest
 
-        # - name: macOS
-        #   os: macos-latest
+        - name: macOS
+          os: macos-latest
 
         - name: Android32
           os: ubuntu-latest
