+++ CMakeLists.txt
@@ -1,13 +1,16 @@
 cmake_minimum_required(VERSION 3.21)
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-set(CMAKE_OSX_ARCHITECTURES "x86_64")
+set(CMAKE_OSX_ARCHITECTURES "arm64;x86_64")
 set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 
-project(OptimizedSongSize VERSION 1.0.0)
+project(SuperExpert VERSION 1.0.0)
 
 add_library(${PROJECT_NAME} SHARED
     src/main.cpp
+    src/ExpertMapLayer.cpp
+    src/ExpertStartupLayer.cpp
+    src/ExpertStartPopup.cpp
 )
 
 if (NOT DEFINED ENV{GEODE_SDK})
