+++ src/utils.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <vector>
+#include <string>
+
+inline std::vector<std::string> splitString(const std::string& s, char delimiter) {
+    std::vector<std::string> tokens;
+    std::istringstream ss(s);
+    std::string token;
+    while (std::getline(ss, token, delimiter)) {
+        tokens.push_back(token); 
+    }
+    return tokens;
+}
